# docs/sales.openapi.yaml
openapi: 3.0.1
info:
  title: sales
  description: ''
  version: 1.0.0
tags: []
paths:
  /token:
    post:
      summary: token
      deprecated: false
      description: ''
      tags: []
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  example: admin
                  type: string
                password:
                  example: '{{key}}'
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties: {}
      security: []
  /sales:
    post:
      summary: delete product
      deprecated: false
      description: ''
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                variables:
                  type: object
                  properties: {}
                  nullable: true
              required:
                - query
            example: >-
              {"query":"mutation {\n    deleteProduct(\n        id: 1\n   
              )\n}\n"}
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties: {}
      security:
        - bearer: []
  /profiles:
    post:
      summary: get role
      deprecated: false
      description: ''
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                variables:
                  type: object
                  properties: {}
                  nullable: true
              required:
                - query
            example: >-
              {"query":"{\n    role(id: 2) { \n        name \n        accounts
              {\n            id\n            name\n        }\n    }\n}"}
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties: {}
      security:
        - bearer: []
components:
  schemas: {}
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
servers: []